writing: transpose technique from software modeling
This is a technique we use for our software modeling group.
It is a precursor to the brain connection towards C4 diagrams
(up for a later PR).

This technique isn't intended to be followed to the letter (unless
you're writing a business essay). This style approach is to encourage
idea outline and converting those outlines into more detailed
explanations with links or idea relations between elaborated concepts.